requirements.txt
slack-bolt==1.23.0
slack-sdk==3.35.0
python-dotenv==1.1.0
requests==2.31.0
